fix: revert broken workspace refactor to restore working state
The workspace service was broken by incomplete refactor in commit 8995094.
That refactor attempted to move everything to internal/ package but:
- Had import cycle issues (noted in commit message as 'wip')
- api.go called non-existent internal.NewManagerImpl()
- Types needed by api.go were moved to internal

Reverting to the working flat package structure so we can:
1. Have a working workspace service again
2. Properly plan a dependable-go refactor if needed
