+++ workspace/pkg/workspace/exec.go
@@ -1,4 +1,4 @@
-package internal
+package workspace
 
 import (
 	"context"
