+++ workspace/pkg/workspace/git_integration_test.go
@@ -1,4 +1,4 @@
-package internal
+package workspace
 
 import (
 	"context"
