+++ workspace/pkg/workspace/workspace.go
@@ -1,4 +1,4 @@
-package internal
+package workspace
 
 import (
 	"time"
