+++ workspace/pkg/workspace/workspace_test.go
@@ -1,4 +1,4 @@
-package internal
+package workspace
 
 import (
 	"testing"
